create project structure
