+++ .gitignore
@@ -17,3 +17,5 @@
 # Ignore logs
 pod_server/logs/log.txt
 relay_server/logs/log.txt
+
+.idea/
